backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
firebase-admin==6.1.0
pydantic==1.10.9
requests==2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
